requirements.txt
# Core dependencies
crewai>=0.28.0
crewai-tools>=0.1.6
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.0.1

# Vector database
chromadb>=0.4.22
sentence-transformers>=2.2.2

# Session management
redis>=5.0.1

# Environment management
python-dotenv>=1.0.0

# Web framework and socket communication
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0

# Additional utilities
pydantic>=2.0.0

python-socketio>=5.0.0
aiohttp>=3.7.0

# Wav2Lip and video processing dependencies
opencv-python>=4.8.0
librosa>=0.10.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
scipy>=1.10.0
moviepy>=1.0.3

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0

# Database
pymongo>=4.6.0

# AWS SDK for Polly TTS
boto3>=1.34.0
